requirements.txt
autopep8==2.3.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
pycodestyle==2.12.1
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
